% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausCor.R
\name{plot_16}
\alias{plot_16}
\title{Save scatter plots}
\usage{
plot_16(a_mat, b_mat, list, out_info)
}
\arguments{
\item{a_mat}{Matrix of measurements of A for each sample.}

\item{b_mat}{Matrix of measurements of B for each sample.}

\item{list}{List of results.}

\item{out_info}{Output directory.}
}
\value{
No return value, saved as a file.
}
\description{
Save scatter plots
}
