% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausCor.R
\name{save_text}
\alias{save_text}
\title{Save list as a text file}
\usage{
save_text(list, out_info, file_type)
}
\arguments{
\item{list}{List of results.}

\item{out_info}{Output directory.}

\item{file_type}{Choose from "excel", "csv", "tsv".}
}
\description{
Save list as a text file
}
