% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausCor.R
\name{filter_n}
\alias{filter_n}
\title{Make list of A-B pair causal correlations - Normal Filtering version}
\usage{
filter_n(a_mat, b_mat, a_category, b_category, min_cor, min_sample, min_r2)
}
\arguments{
\item{a_mat}{Matrix of measurements of A for each sample.}

\item{b_mat}{Matrix of measurements of B for each sample.}

\item{a_category}{Category name of A.}

\item{b_category}{Category name of B.}

\item{min_cor}{Minimum spearman correlation coefficient.}

\item{min_sample}{Minimum number of samples.}

\item{min_r2}{Minimum R2.}
}
\description{
Make list of A-B pair causal correlations - Normal Filtering version
}
