% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausCor.R
\name{plot_16}
\alias{plot_16}
\title{Save scatter plots}
\usage{
plot_16(a_mat, b_mat, list, out_info, x_italic = F, y_italic = T)
}
\arguments{
\item{a_mat}{Matrix of measurements of A for each sample.}

\item{b_mat}{Matrix of measurements of B for each sample.}

\item{list}{List of results.}

\item{out_info}{Output directory.}

\item{x_italic}{Italicize the x-axis label of the plot. The default is False.}

\item{y_italic}{Italicize the y-axis label of the plot. The default is True.}
}
\description{
Save scatter plots
}
